task: move most functions from a header to a new .c file
Move large functions from src/arch/xtensa/include/arch/task.h to
src/arch/xtensa/smp/init.c to avoid building them multiple times.
